PycharmProjects/sports_funder/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# AI & External APIs
google-generativeai==0.3.2
twilio==8.10.3
requests==2.31.0

# Payment Processing
stripe==7.8.0
paypalrestsdk==1.13.3
squareup==4.0.0

# QR Code Generation
qrcode[pil]==7.4.2

# Image Processing
Pillow==10.1.0

# Cloud Storage
google-cloud-storage==2.10.0
google-cloud-sql-connector==1.4.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Environment Management
python-dotenv==1.0.0

